Helm charts V1.1, plus Makefile and test improvements

+++ install-mq.sh
@@ -18,25 +18,53 @@
 # Fail on any non-zero return code
 set -ex
 
-export DEBIAN_FRONTEND=noninteractive
+test -f /usr/bin/yum && RHEL=true || RHEL=false
+test -f /usr/bin/apt-get && UBUNTU=true || UBUNTU=false
+
+# If MQ_PACKAGES isn't specifically set, then choose a valid set of defaults
+if [ -z $MQ_PACKAGES ]; then
+  $UBUNTU && MQ_PACKAGES="ibmmq-server ibmmq-java ibmmq-jre ibmmq-gskit ibmmq-msg-.* ibmmq-samples ibmmq-ams"
+  $RHEL && MQ_PACKAGES="MQSeriesRuntime-*.rpm MQSeriesServer-*.rpm MQSeriesJava*.rpm MQSeriesJRE*.rpm MQSeriesGSKit*.rpm MQSeriesMsg*.rpm MQSeriesSamples*.rpm MQSeriesAMS-*.rpm"
+fi
+
+if ($UBUNTU); then
+  export DEBIAN_FRONTEND=noninteractive
+  # Install additional packages required by MQ, this install process and the runtime scripts
+  apt-get update
+  apt-get install -y --no-install-recommends \
+    bash \
+    bc \
+    ca-certificates \
+    coreutils \
+    curl \
+    debianutils \
+    file \
+    findutils \
+    gawk \
+    grep \
+    libc-bin \
+    mount \
+    passwd \
+    procps \
+    sed \
+    tar \
+    util-linux
+fi
 
 # Install additional packages required by MQ, this install process and the runtime scripts
-apt-get update
-apt-get install -y --no-install-recommends \
+$RHEL && yum -y install \
   bash \
   bc \
   ca-certificates \
   coreutils \
   curl \
-  debianutils \
   file \
   findutils \
   gawk \
+  glibc-common \
   grep \
-  libc-bin \
-  mount \
   passwd \
-  procps \
+  procps-ng \
   sed \
   tar \
   util-linux
@@ -49,30 +77,37 @@ curl -LO $MQ_URL
 tar -zxvf ./*.tar.gz
 
 # Remove packages only needed by this script
-apt-get purge -y \
+$UBUNTU && apt-get purge -y \
   ca-certificates \
   curl
 
+# Note: ca-certificates and curl are installed by default in RHEL
+
 # Remove any orphaned packages
-apt-get autoremove -y
+$UBUNTU && apt-get autoremove -y
 
 # Recommended: Create the mqm user ID with a fixed UID and group, so that the file permissions work between different images
-groupadd --system --gid 999 mqm
-useradd --system --uid 999 --gid mqm mqm
+$UBUNTU && groupadd --system --gid 999 mqm
+$UBUNTU && useradd --system --uid 999 --gid mqm mqm
+$RHEL && groupadd --system --gid 888 mqm
+$RHEL && useradd --system --uid 888 --gid mqm mqm
 usermod -G mqm root
 
 # Find directory containing .deb files
-DIR_DEB=$(find ${DIR_EXTRACT} -name "*.deb" -printf "%h\n" | sort -u | head -1)
+$UBUNTU && DIR_DEB=$(find ${DIR_EXTRACT} -name "*.deb" -printf "%h\n" | sort -u | head -1)
+$RHEL && DIR_RPM=$(find ${DIR_EXTRACT} -name "*.rpm" -printf "%h\n" | sort -u | head -1)
 # Find location of mqlicense.sh
 MQLICENSE=$(find ${DIR_EXTRACT} -name "mqlicense.sh")
 
 # Accept the MQ license
 ${MQLICENSE} -text_only -accept
-echo "deb [trusted=yes] file:${DIR_DEB} ./" > /etc/apt/sources.list.d/IBM_MQ.list
+$UBUNTU && echo "deb [trusted=yes] file:${DIR_DEB} ./" > /etc/apt/sources.list.d/IBM_MQ.list
 
 # Install MQ using the DEB packages
-apt-get update
-apt-get install -y $MQ_PACKAGES
+$UBUNTU && apt-get update
+$UBUNTU && apt-get install -y $MQ_PACKAGES
+
+$RHEL && cd $DIR_RPM && rpm -ivh $MQ_PACKAGES
 
 # Remove 32-bit libraries from 64-bit container
 find /opt/mqm /var/mqm -type f -exec file {} \; | awk -F: '/ELF 32-bit/{print $1}' | xargs --no-run-if-empty rm -f
@@ -84,19 +119,21 @@ find /opt/mqm -name '*.tar.gz' -delete
 /opt/mqm/bin/setmqinst -p /opt/mqm -i
 
 # Clean up all the downloaded files
-rm -f /etc/apt/sources.list.d/IBM_MQ.list
+$UBUNTU && rm -f /etc/apt/sources.list.d/IBM_MQ.list
 rm -rf ${DIR_EXTRACT}
 
-#### Apply any bug fixes not included in base Ubuntu or MQ image.
+# Apply any bug fixes not included in base Ubuntu or MQ image.
 # Don't upgrade everything based on Docker best practices https://docs.docker.com/engine/userguide/eng-image/dockerfile_best-practices/#run
-apt-get upgrade -y libdb5.3
-#### End of bug fixes
 
-# Clean up cached apt files
-rm -rf /var/lib/apt/lists/*
+# End of bug fixes
+
+# Clean up cached files
+$UBUNTU && rm -rf /var/lib/apt/lists/*
+$RHEL && yum -y clean all
+$RHEL && rm -rf /var/cache/yum/*
 
 # Optional: Update the command prompt with the MQ version
-echo "mq:$(dspmqver -b -f 2)" > /etc/debian_chroot
+$UBUNTU && echo "mq:$(dspmqver -b -f 2)" > /etc/debian_chroot
 
 # Remove the directory structure under /var/mqm which was created by the installer
 rm -rf /var/mqm
@@ -113,4 +150,7 @@ ln -s /mnt/mqm/data /var/mqm
 # Optional: Set these values for the Bluemix Vulnerability Report
 sed -i 's/PASS_MAX_DAYS\t99999/PASS_MAX_DAYS\t90/' /etc/login.defs
 sed -i 's/PASS_MIN_DAYS\t0/PASS_MIN_DAYS\t1/' /etc/login.defs
-sed -i 's/password\t\[success=1 default=ignore\]\tpam_unix\.so obscure sha512/password\t[success=1 default=ignore]\tpam_unix.so obscure sha512 minlen=8/' /etc/pam.d/common-password
+
+$UBUNTU && PAM_FILE=/etc/pam.d/common-password
+$RHEL && PAM_FILE=/etc/pam.d/password-auth
+sed -i 's/password\t\[success=1 default=ignore\]\tpam_unix\.so obscure sha512/password\t[success=1 default=ignore]\tpam_unix.so obscure sha512 minlen=8/' $PAM_FILE
